requirements.txt
requests==2.22.0
behave==1.2.6
PyHamcrest==2.0.0
jsonschema==3.2.0